* magick/studio.h (_FILE_OFFSET_BITS): Fix _FILE_OFFSET_BITS #ifdef to avoid warnings when it is not defined.
* configure.ac: Fix minor shell syntax error (used ==).


* Merged in updates from 1.1 release branch.
* version.sh (PACKAGE_VERSION): Release version 1.1.
(PACKAGE_VERSION): Next major release will be 1.1.  Bump library
versions since we anticipate adding interfaces.
* magick/studio.h: Disabled use of Windows message lookups
entirely since this doesn't seem to work for programs.  It
probably only works for DLLs like ImageMagickObject.
* magick/delegate.c (ListDelegateInfo): Don't get stuck in a loop
if fprintf() returns zero.


* VisualMagick/*/*.bat: Add .exe to exe file names in order to
ensure that the executable is executed rather than something else
(such as a batch script).
* magick/locale.c: Only use Windows resource message lookup for
the DLL build since it doesn't seem to work for static
executables.
* magick/utility.c (SetClientPath): Ensure that client path is
null terminated.
(SetClientFilename): Initialize default client filename to ""
rather than "gm.exe" and ensure that filename is null terminated.
(SetClientName): Ensure that client path is null terminated.
* magick/blob.c (GetConfigureBlob): Enable logging the load of
log.mgk
* magick/log.c: Re-designed logging initialization in order to
avoid the "chicken and the egg" snafu.  This allows logging the
initialization of the logging subsystem.


* VisualMagick/bin/log.mgk: Add comments to help document usage.
* config/log.mgk: Add comments to help document usage.
* magick/utility.c (IsAccessible): Use access() rather than stat().
(IsAccessibleNoLogging): Use access() rather than stat().
(IsDirectory): Implemented return status according to existing API
documentation.
(GetExecutionPathUsingName): Complete re-write in order to minimize
path computation logic and fix failure with partial paths.


* doc/options.imdoc: Add "Color" to the arguments available for -list.


* version.sh (PACKAGE_CHANGE_DATE): Updated version to 1.1Beta3
* magick/nt_base.c (NTSystemComman): Have not been successful with using
MsgWaitForMultipleObjects() reliably, so back out usage of it for
now.
* magick/fx.c (ConvolveImage): Support convolution in CMYK
colorspace.


* version.sh (PACKAGE_CHANGE_DATE): Update release to 1.1Beta2.
* magick/constitute.c (ReadImage): Ignore errors of type
ConfigureError when loading modules so that error report can still
report the familiar "No delegate for this image format" message
while still being able to report problems with loading a module if
it was found.  This is a compromise which reports useless messages
when the modules are not found at all, but I can't think of a way
around it.
* magick/nt_base.c (NTSystemComman): Adjust
MsgWaitForMultipleObjects() arguments to wait for object to be
signaled only (FALSE) rather wait for object to be signaled as
well as an input event received(TRUE).  It seems that process
status changes do not result in an input event, so the call was
deadlocking.
* magick/constitute.c (ReadImage): If module loading reported an
error, ReadImage immediately returned an error rather than trying
to use a delegate defined by delegates.mgk.


* version.sh: Updated version to 1.1Beta1