* magick/pixel_cache.c: Exhuastive study of the pixel cache code reveals that it is inscrutable and not implemented in a fashion which enables useful multi-threading.  Therefore, the cache view interfaces are now made OpenMP-safe via a global critical section.
* magick/pixel_iterator.c: Reduce use of critical sections around
pixel cache to the bare minimum based on analysis and testing.
Unfortunately, testing shows that the pixel cache views are still
not 100% thread safe so the extra locking is still required.
* magick/pixel_cache.c (ModifyCache): Make implementation thread
safe.  This required removing a thread-unsafe optimization from
Bill Radcliffe.
* magick/command.c (BenchmarkImageCommand): Restore original
client name for each loop so that it is not extended further for
each iteration.
* magick/semaphore.c (UnlockSemaphoreInfo): Decrement lock depth
under protection of the lock.