SVG: reenable use of '>' in geometry strings for SVG files
A terminating '>' in a geometry string is interpreted to mean that
the dimensions of an image should only be changed if its width or
height exceeds the geometry specification.  For an unapparent and
undocumented reason, a terminating '>', if present, was being nulled
out, making this feature unusable for SVG files (now fixed).