SVG: Clipping path is a geometric silhouette.
According to the SVG spec, a clipping path is defined only
by the geometry of its constituent elements, and is not
dependent on fill color/opacity, stroke color/opacity, or
stroke width.  To ensure conformity with the spec, when a
clipping path is created, these SVG elements are set to
appropriate values, and any attempt to modify them is ignored.

Also, whenever a clipping path is drawn, the associated image
attributes are now updated from the parent image structure.
This ensures that any added or modified attributes are up to
date.