SVG: Properly track current text position (+ other).
The current text position is now maintained by DrawImage()
(render.c) instead of by SVGStartElement() and SVGEndElement()
(svg.c).  This change was made to support the recently
implmemented "use" and "class" elements, which may make
changes to the font size that are not visible to the code
in svg.c.

When processing SVG attributes associated with an SVG
element, the attribute list is reordered so that the
"font-size", "class", and "style" attributes are always
processed first.  This ensures that a change to the font
size will be processed before any dimensional attribute
whose value may depend on the font size (e.g., a width
value specified in "em" units).

Fixed two memory leaks in ProcessStyleClassDefs() (svg.c)
associated with making an early return when malformed
input is detected.

Fixed an uninitialized variable condition in
ExtractTokensBetweenPushPop() (render.c) which can occur
when malformed input is detected.

DrawInfo member "unused1" member variable has been renamed
"flags".  It is now used to tag a DrawInfo as being a
clipping path or a compositing mask.